refactor: flow getting value of app from context
