+++ src/components/Providers.tsx
@@ -5,11 +5,11 @@ import { ChakraProvider, ColorModeScript } from "@chakra-ui/react";
 import { MDXProvider } from "@mdx-js/react";
 
 import { AppProvider } from "@/contexts/AppContext/AppContext";
+import { App } from "@/generated/graphql";
 // import { ConfigurationContextProvider } from "@/contexts/ConfigurationContext/ConfigurationContext";
 import { ApolloProvider } from "@/libs/apollo/client";
 import { theme } from "@/theme";
 import { MDXComponents } from "@/theme/MDXComponents";
-import { App } from "@/types/Hygraph/models/App";
 import { Configuration } from "@/types/Hygraph/models/Configuration";
 
 type ProvidersProps = React.HTMLProps<HTMLElement> & {
