+++ src/contexts/AppContext/AppContext.types.tsx
@@ -1,4 +1,4 @@
-import { App } from "@/types/Hygraph/models/App";
+import { App } from "@/generated/graphql";
 
 export enum PROJECT_ACTION {}
 
@@ -7,10 +7,10 @@ export type AppAction = {
   payload?: unknown;
 };
 
-export type AppState = App;
+export type AppState = Partial<App>;
 
 export type AppContext = {
-  apps: Array<App>;
+  apps: Partial<App>[];
 };
 
 export type AppProviderProps = React.ComponentPropsWithoutRef<"div"> & {
